// src/generated/models/JointPoleIntentsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface ProcessStageValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface Partner_VendorValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface StatusValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface AuthorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface EditorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface JointPoleIntentsBase {
  ID?: number;
  Title?: string;
  IntentNumber0?: string;
  IntentNumber?: string;
  IntentLifecycleStage?: string;
  Description?: string;
  F2Description?: string;
  "ProcessStage#Id"?: number;
  "Partner_Vendor#Id"?: number;
  PoleID?: string;
  LocationCity?: string;
  LocationState?: string;
  "Status#Id"?: number;
  ApprovalDate?: string;
  OData__ColorTag?: string;
  ComplianceAssetId?: string;
  Modified?: string;
  Created?: string;
  "Author#Claims"?: string;
  "Editor#Claims"?: string;
  "{Identifier}"?: string;
  "{IsFolder}"?: boolean;
  "{Thumbnail}"?: {
    "@odata.type"?: string;
    // URL to the small, highly compressed size thumbnail of the item, if available.
    Small?: string;
    // URL to the standard size thumbnail of the item, if available.
    Medium?: string;
    // URL to the largest size thumbnail of the item, if available.
    Large?: string;
  };
  "{Link}"?: string;
  "{Name}"?: string;
  "{FilenameWithExtension}"?: string;
  "{Path}"?: string;
  "{FullPath}"?: string;
  "{ModerationStatus}"?: string;
  "{ModerationComment}"?: string;
  "{ContentType}#Id"?: string;
  "{ContentType}"?: {
    "@odata.type"?: string;
    Id?: string;
    Name?: string;
  };
  "{HasAttachments}"?: boolean;
  "{Attachments}@odata.type"?: string;
  "{Attachments}"?: {
   "@odata.type"?: string;
   Id?: string;
   AbsoluteUri?: string;
   DisplayName?: string;
 }[];
  "{VersionNumber}"?: string;
  "{TriggerWindowStartToken}"?: string;
  "{TriggerWindowEndToken}"?: string;
}

export interface JointPoleIntentsRead extends JointPoleIntentsBase {
  ProcessStage?: ProcessStageValue;
  Partner_Vendor?: Partner_VendorValue;
  Status?: StatusValue;
  Author?: AuthorValue;
  Editor?: EditorValue;
}

export interface JointPoleIntentsWrite extends JointPoleIntentsBase {
  ProcessStage?: string;
  Partner_Vendor?: string;
  Status?: string;
  Author?: AuthorValue;
  Editor?: EditorValue;
}